pyproject.toml
[project]
name = "learnit-oxford-translation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "openai>=1.59.9",
    "pyjwt>=2.10.1",
]
